Avance del consultar usuario
Vista y procedimiento listo. Swagger listo, falta funcionalidad desde web

+++ ProyectoAPI/ProyectoAPI/Entities/UsuarioEnt.cs
@@ -8,7 +8,7 @@
         public string CorreoElectronico { get; set; } = string.Empty;
         public string Contrasenna { get; set; } = string.Empty;
         public bool Estado { get; set; }
-        public Byte IdRol { get; set; }
+        public long IdRol { get; set; }
 
         public string contrasennaTemporal { get; set; } = string.Empty;
